init 修改建表sql（bigint——>int）

+++ script/create_tables.sql
@@ -1,6 +1,6 @@
 -- 老师表
 CREATE TABLE teachers (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT,
+    id INT PRIMARY KEY AUTO_INCREMENT,
     teacher_code VARCHAR(50) NOT NULL UNIQUE COMMENT '老师业务编号',
     name VARCHAR(100) NOT NULL COMMENT '姓名',
     phone VARCHAR(20) COMMENT '联系电话',
@@ -12,10 +12,10 @@ CREATE TABLE teachers (
 
 -- 班级表
 CREATE TABLE classes (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT,
+    id INT PRIMARY KEY AUTO_INCREMENT,
     class_code VARCHAR(50) NOT NULL UNIQUE COMMENT '班级业务编号',
     start_date DATE NOT NULL COMMENT '开课时间',
-    head_teacher_id BIGINT COMMENT '班主任ID (逻辑关联 teachers.id)',
+    head_teacher_id INT COMMENT '班主任ID (逻辑关联 teachers.id)',
     is_deleted TINYINT(1) DEFAULT 0 COMMENT '逻辑删除标志',
     create_time DATETIME DEFAULT CURRENT_TIMESTAMP,
     update_time DATETIME DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
@@ -23,9 +23,9 @@ CREATE TABLE classes (
 
 -- 班级授课周期表
 CREATE TABLE class_teaching_periods (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT,
-    class_id BIGINT NOT NULL COMMENT '班级ID (逻辑关联 classes.id)',
-    lecturer_id BIGINT NOT NULL COMMENT '授课老师ID (逻辑关联 teachers.id)',
+    id INT PRIMARY KEY AUTO_INCREMENT,
+    class_id INT NOT NULL COMMENT '班级ID (逻辑关联 classes.id)',
+    lecturer_id INT NOT NULL COMMENT '授课老师ID (逻辑关联 teachers.id)',
     start_date DATE NOT NULL COMMENT '授课开始日期',
     end_date DATE COMMENT '授课结束日期 (NULL表示当前仍在授课)',
     is_deleted TINYINT(1) DEFAULT 0 COMMENT '逻辑删除标志',
@@ -35,10 +35,10 @@ CREATE TABLE class_teaching_periods (
 
 -- 学生表
 CREATE TABLE students (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT,
+    id INT PRIMARY KEY AUTO_INCREMENT,
     student_code VARCHAR(50) NOT NULL UNIQUE COMMENT '学生业务编号',
-    class_id BIGINT NOT NULL COMMENT '所属班级ID (逻辑关联 classes.id)',
-    advisor_id BIGINT COMMENT '顾问老师ID (逻辑关联 teachers.id)',
+    class_id INT NOT NULL COMMENT '所属班级ID (逻辑关联 classes.id)',
+    advisor_id INT COMMENT '顾问老师ID (逻辑关联 teachers.id)',
     name VARCHAR(100) NOT NULL COMMENT '姓名',
     gender TINYINT NOT NULL COMMENT '性别: 0-女, 1-男',
     age INT COMMENT '年龄',
@@ -55,8 +55,8 @@ CREATE TABLE students (
 
 -- 成绩表
 CREATE TABLE scores (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT,
-    student_id BIGINT NOT NULL COMMENT '学生ID (逻辑关联 students.id)',
+    id INT PRIMARY KEY AUTO_INCREMENT,
+    student_id INT NOT NULL COMMENT '学生ID (逻辑关联 students.id)',
     exam_sequence VARCHAR(20) NOT NULL COMMENT '考核序次 (如: 期中, 期末, 阶段1)',
     score DECIMAL(5,2) NOT NULL COMMENT '成绩 (0.00-100.00)',
     is_deleted TINYINT(1) DEFAULT 0 COMMENT '逻辑删除标志',
@@ -67,8 +67,8 @@ CREATE TABLE scores (
 
 -- 就业表
 CREATE TABLE employments (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT,
-    student_id BIGINT NOT NULL COMMENT '学生ID (逻辑关联 students.id)',
+    id INT PRIMARY KEY AUTO_INCREMENT,
+    student_id INT NOT NULL COMMENT '学生ID (逻辑关联 students.id)',
     company_name VARCHAR(200) NOT NULL COMMENT '就业公司名称',
     job_open_date DATETIME NOT NULL COMMENT '就业开放时间',
     offer_date DATETIME COMMENT 'offer下发时间',
